pipeline/requirements.txt
httpx>=0.27
anthropic>=0.40
google-genai>=1.0
beautifulsoup4>=4.12
Pillow>=10.0
